requirements/base.txt
APScheduler==3.7.0
requests==2.25.1
lxml==4.6.3
PyExecJS==1.5.1
numpy==1.19.5
tensorflow-cpu==2.5.0
yagmail==0.14.256
Pillow==8.2.0
opencv-python==4.5.2.52

